Revert "chore: fix performance-enum-size warning in gtk client"
This reverts commit 5ce6b562f849c2499fa34f4d903234f1945f9c3e.

+++ gtk/Torrent.cc
@@ -116,7 +116,7 @@ Torrent::Columns::Columns()
 class Torrent::Impl
 {
 public:
-    enum class Property : guint8
+    enum class Property : guint
     {
         ICON = 1,
         NAME,
